refactor(recipe-scraper): Extract shared input state/UI from the demo and utility scrapers.
RecipeScraperDemo and RecipeScraperUtilityContent had nearly identical
url/text/image state, mutual-exclusivity handlers, and input JSX
copy-pasted between them, differing only in what happens after the
scrape (preview vs. save-to-file). Extracts a useRecipeScraperInputs
hook and a RecipeScraperInputs component so both consumers share one
implementation, resolving their complexity lint warnings as a result.
