+++ projects/nx-workspace/apps/ui/vb-manager-next/src/app/components/llm/RecipeScraperDemo.tsx
@@ -1,34 +1,15 @@
 'use client';
 import { HTTP_METHOD, HTTP_HEADERS } from '@vigilant-broccoli/common-js';
 import { Text } from '@radix-ui/themes';
-import {
-  Button,
-  CopyPastable,
-  Input,
-  Textarea,
-} from '@vigilant-broccoli/react-lib';
+import { Button, CopyPastable } from '@vigilant-broccoli/react-lib';
 import { useState } from 'react';
 import { API_ENDPOINTS } from '../../constants/api-endpoints';
 import { authFetch } from '../../../../libs/auth';
-import {
-  UploadedImage,
-  readImageAsBase64,
-} from '../../utils/image-upload.utils';
-
-const buildScrapeRequestBody = (
-  recipeUrl: string,
-  recipeText: string,
-  image: UploadedImage | null,
-) => {
-  if (image) return { images: [image] };
-  if (recipeText.trim()) return { text: recipeText };
-  return { url: recipeUrl };
-};
+import { useRecipeScraperInputs } from '../../hooks/useRecipeScraperInputs';
+import { RecipeScraperInputs } from './RecipeScraperInputs';
 
 export const RecipeScraperDemo = () => {
-  const [recipeUrl, setRecipeUrl] = useState('');
-  const [recipeText, setRecipeText] = useState('');
-  const [image, setImage] = useState<UploadedImage | null>(null);
+  const inputs = useRecipeScraperInputs();
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [result, setResult] = useState<{
@@ -36,28 +17,8 @@ export const RecipeScraperDemo = () => {
     markdown: string;
   } | null>(null);
 
-  const handleFileSelect = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setImage(await readImageAsBase64(file));
-    setRecipeUrl('');
-    setRecipeText('');
-  };
-
-  const clearImage = () => setImage(null);
-
-  const handleUrlChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setRecipeUrl(e.target.value);
-    if (e.target.value) setRecipeText('');
-  };
-
-  const handleTextChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setRecipeText(e.target.value);
-    if (e.target.value) setRecipeUrl('');
-  };
-
   const handleScrapeRecipe = async () => {
-    if (!recipeUrl.trim() && !recipeText.trim() && !image) {
+    if (!inputs.hasAnyInput) {
       setError('Enter a URL, paste recipe text, or select an image');
       return;
     }
@@ -71,9 +32,7 @@ export const RecipeScraperDemo = () => {
       headers: {
         ...HTTP_HEADERS.CONTENT_TYPE.JSON,
       },
-      body: JSON.stringify(
-        buildScrapeRequestBody(recipeUrl, recipeText, image),
-      ),
+      body: JSON.stringify(inputs.requestBody),
     });
 
     const data = await response.json();
@@ -94,51 +53,17 @@ export const RecipeScraperDemo = () => {
     }
   };
 
-  const hasUrl = !!recipeUrl.trim();
-  const hasText = !!recipeText.trim();
-  const hasImage = !!image;
-  const hasAnyInput = hasUrl || hasText || hasImage;
-
   return (
     <div className="flex flex-col gap-3">
-      <Input
-        placeholder="Enter recipe URL..."
-        value={recipeUrl}
-        onChange={handleUrlChange}
-        onKeyDown={handleRecipeKeyDown}
-        disabled={loading || hasImage || hasText}
+      <RecipeScraperInputs
+        inputs={inputs}
+        loading={loading}
+        onUrlKeyDown={handleRecipeKeyDown}
       />
-
-      <Textarea
-        placeholder="...or paste recipe text here"
-        value={recipeText}
-        onChange={handleTextChange}
-        disabled={loading || hasImage || hasUrl}
-        rows={6}
-      />
-
-      <div className="flex gap-2 items-center">
-        <input
-          type="file"
-          accept="image/*"
-          onChange={handleFileSelect}
-          disabled={loading || hasUrl || hasText}
-        />
-        {image && (
-          <>
-            <Text size="2" color="gray">
-              {image.name}
-            </Text>
-            <Button onClick={clearImage} disabled={loading} variant="outline">
-              Clear
-            </Button>
-          </>
-        )}
-      </div>
 
       <Button
         onClick={handleScrapeRecipe}
-        disabled={loading || !hasAnyInput}
+        disabled={loading || !inputs.hasAnyInput}
         className="self-end"
       >
         {loading ? 'Scraping...' : 'Scrape'}
