+++ projects/nx-workspace/apps/ui/vb-manager-next/src/app/components/utilities/recipe-scraper.utility.tsx
@@ -1,67 +1,28 @@
 'use client';
 import { HTTP_METHOD, HTTP_HEADERS } from '@vigilant-broccoli/common-js';
 import { Text } from '@radix-ui/themes';
-import { Button, Input, Textarea } from '@vigilant-broccoli/react-lib';
+import { Button } from '@vigilant-broccoli/react-lib';
 import { useState } from 'react';
 import { API_ENDPOINTS } from '../../constants/api-endpoints';
 import { authFetch } from '../../../../libs/auth';
-import {
-  UploadedImage,
-  readImageAsBase64,
-} from '../../utils/image-upload.utils';
+import { useRecipeScraperInputs } from '../../hooks/useRecipeScraperInputs';
+import { RecipeScraperInputs } from '../llm/RecipeScraperInputs';
 
 const MESSAGE_TYPE = {
   SUCCESS: 'success',
   ERROR: 'error',
 } as const;
 
-const buildScrapeRequestBody = (
-  recipeUrl: string,
-  recipeText: string,
-  image: UploadedImage | null,
-) => {
-  if (image) return { images: [image] };
-  if (recipeText.trim()) return { text: recipeText };
-  return { url: recipeUrl };
-};
-
 export const RecipeScraperUtilityContent = () => {
-  const [recipeUrl, setRecipeUrl] = useState('');
-  const [recipeText, setRecipeText] = useState('');
-  const [image, setImage] = useState<UploadedImage | null>(null);
+  const inputs = useRecipeScraperInputs();
   const [recipeLoading, setRecipeLoading] = useState(false);
   const [recipeMessage, setRecipeMessage] = useState<{
     type: (typeof MESSAGE_TYPE)[keyof typeof MESSAGE_TYPE];
     text: string;
   } | null>(null);
 
-  const hasUrl = !!recipeUrl.trim();
-  const hasText = !!recipeText.trim();
-  const hasImage = !!image;
-  const hasAnyInput = hasUrl || hasText || hasImage;
-
-  const handleFileSelect = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setImage(await readImageAsBase64(file));
-    setRecipeUrl('');
-    setRecipeText('');
-  };
-
-  const clearImage = () => setImage(null);
-
-  const handleUrlChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setRecipeUrl(e.target.value);
-    if (e.target.value) setRecipeText('');
-  };
-
-  const handleTextChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setRecipeText(e.target.value);
-    if (e.target.value) setRecipeUrl('');
-  };
-
   const handleScrapeRecipe = async () => {
-    if (!hasAnyInput) {
+    if (!inputs.hasAnyInput) {
       setRecipeMessage({
         type: MESSAGE_TYPE.ERROR,
         text: 'Enter a URL, paste recipe text, or select an image',
@@ -77,9 +38,7 @@ export const RecipeScraperUtilityContent = () => {
       headers: {
         ...HTTP_HEADERS.CONTENT_TYPE.JSON,
       },
-      body: JSON.stringify(
-        buildScrapeRequestBody(recipeUrl, recipeText, image),
-      ),
+      body: JSON.stringify(inputs.requestBody),
     });
 
     if (!response.ok) {
@@ -91,9 +50,7 @@ export const RecipeScraperUtilityContent = () => {
       setRecipeLoading(false);
       return;
     }
-    setRecipeUrl('');
-    setRecipeText('');
-    setImage(null);
+    inputs.reset();
     setRecipeMessage({
       type: MESSAGE_TYPE.SUCCESS,
       text: 'Recipe downloaded successfully!',
@@ -110,48 +67,15 @@ export const RecipeScraperUtilityContent = () => {
   return (
     <>
       <div className="flex flex-col gap-3">
-        <Input
-          placeholder="Enter recipe URL..."
-          value={recipeUrl}
-          onChange={handleUrlChange}
-          onKeyDown={handleRecipeKeyDown}
-          disabled={recipeLoading || hasImage || hasText}
+        <RecipeScraperInputs
+          inputs={inputs}
+          loading={recipeLoading}
+          onUrlKeyDown={handleRecipeKeyDown}
         />
-
-        <Textarea
-          placeholder="...or paste recipe text here"
-          value={recipeText}
-          onChange={handleTextChange}
-          disabled={recipeLoading || hasImage || hasUrl}
-          rows={6}
-        />
-
-        <div className="flex gap-2 items-center">
-          <input
-            type="file"
-            accept="image/*"
-            onChange={handleFileSelect}
-            disabled={recipeLoading || hasUrl || hasText}
-          />
-          {image && (
-            <>
-              <Text size="2" color="gray">
-                {image.name}
-              </Text>
-              <Button
-                onClick={clearImage}
-                disabled={recipeLoading}
-                variant="outline"
-              >
-                Clear
-              </Button>
-            </>
-          )}
-        </div>
 
         <Button
           onClick={handleScrapeRecipe}
-          disabled={recipeLoading || !hasAnyInput}
+          disabled={recipeLoading || !inputs.hasAnyInput}
           className="self-end"
         >
           {recipeLoading ? 'Scraping...' : 'Download'}
